requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.0
pathtools==0.1.2
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
schedule==0.6.0
six==1.14.0
urllib3==1.25.9
webdavclient3==3.14.3
